Create Text for Docker Compose
This is the actual text for the docker compose file.  The OpenVPN or Wireguard info needs to be filled out and whichever isn't used need to be turned into a comment. The containers run as a full stack and require gluetun to be running and functioning in order to work. This is my stack, but the inspiration came from a few places. Also I am not using this anymore as I broke it up so that watchtower can apply updates to individual containers.